chore: topbar, navbar 변경사항 적용
